+++ src/dmqclib/common/config/yaml_templates.py
@@ -81,7 +81,7 @@ feature_param_sets:
         stats_set: { type: raw }
         col_names: [ longitude, latitude ]
       - feature: day_of_year
-        convert: sine
+        convert: cosine
         col_names: [ profile_timestamp ]
       - feature: profile_summary_stats
         stats_set: { type: raw }
@@ -111,7 +111,7 @@ feature_param_sets:
         stats_set: { type: min_max, name: location }
         col_names: [ longitude, latitude ]
       - feature: day_of_year
-        convert: sine
+        convert: cosine
         col_names: [ profile_timestamp ]
       - feature: profile_summary_stats
         stats_set: { type: min_max, name: profile_summary_stats }