+++ tests/test_prepare_features.py
@@ -251,6 +251,31 @@ class TestDayOfYearFeature(_TestFeatureBase):
         ds.scale_second()
         self.assertTrue(ds.features.equals(features))
 
+    def test_convert_cosine(self):
+        """
+        Verifies that day-of-year features are correctly extracted and scaled,
+        specifically testing with sine conversion, and checking the resulting
+        DataFrame type and dimensions.
+        """
+        ds = DayOfYearFeat(
+            "temp",
+            self.feature_info,
+            self.ds_select.selected_profiles,
+            self.ds_extract.filtered_input,
+            self.ds_locate.selected_rows,
+            self.ds_summary.summary_stats,
+        )
+        ds.feature_info = {
+            "class": "day_of_year",
+            "convert": "cosine",
+        }
+        ds.extract_features()
+        ds.scale_second()
+
+        self.assertIsInstance(ds.features, pl.DataFrame)
+        self.assertEqual(ds.features.shape[0], 128)
+        self.assertEqual(ds.features.shape[1], 2)
+
 
 class TestProfileSummaryStatsFeature(_TestFeatureBase):
     """